Align spec plugin with SDK-CLI-Plugin pattern
All plugin commands now delegate to named agents (noun form) which call the CLI.
Command hierarchy: Plugin Command → Agent → CLI → SDK

CLI changes:
- Reorganized spec/index.ts to be thin index pattern
- Created spec/spec.ts with 7 flat dash commands (spec-create, spec-get, spec-list, spec-update, spec-delete, spec-validate, spec-refine)
- Created spec/template.ts with template-list command
- All commands renamed from nested (create, get, ...) to flat dashes (spec-create, spec-get, ...)

Plugin commands (9 total):
- Deleted 4 old commands (create.md, validate.md, refine.md, archive.md)
- Created 9 new commands delegating to agents via Task:
  spec-create, spec-get, spec-list, spec-update, spec-delete, spec-validate, spec-refine, spec-archive, template-list

Plugin agents (9 total):
- Renamed 4 existing agents to noun form (spec-creator, spec-validator, spec-refiner, spec-archiver)
- Created 5 new thin agents for CRUD operations (spec-getter, spec-lister, spec-updater, spec-deleter, template-lister)

Plugin config:
- Updated plugin.json agents array to list all 9 agents

Verified:
- TypeScript build passes cleanly
- fractary-core spec --help shows all 8 flat dash-separated commands
