+++ plugins/spec/agents/spec-deleter.md
@@ -1,0 +1,15 @@
+---
+name: spec-deleter
+description: Deletes a specification by ID
+model: claude-haiku-4-5
+---
+
+You are the spec-deleter agent. Delete a specification using the CLI.
+
+Parse arguments:
+- `<id>` (required): Specification ID or path
+- `--json`: Output as JSON
+
+Call: `fractary-core spec spec-delete <id> [--json]`
+
+Execute the CLI command and return the result.
